+++ perception/radar_object_tracker/include/autoware_radar_object_tracker/association/solver/ssp.hpp
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__RADAR_OBJECT_TRACKER__ASSOCIATION__SOLVER__SSP_HPP_
-#define AUTOWARE__RADAR_OBJECT_TRACKER__ASSOCIATION__SOLVER__SSP_HPP_
+#ifndef AUTOWARE_RADAR_OBJECT_TRACKER__ASSOCIATION__SOLVER__SSP_HPP_
+#define AUTOWARE_RADAR_OBJECT_TRACKER__ASSOCIATION__SOLVER__SSP_HPP_
 
-#include "autoware/radar_object_tracker/association/solver/gnn_solver_interface.hpp"
+#include "autoware_radar_object_tracker/association/solver/gnn_solver_interface.hpp"
 
 #include <unordered_map>
 #include <vector>
@@ -34,4 +34,4 @@ public:
 };
 }  // namespace autoware::radar_object_tracker::gnn_solver
 
-#endif  // AUTOWARE__RADAR_OBJECT_TRACKER__ASSOCIATION__SOLVER__SSP_HPP_
+#endif  // AUTOWARE_RADAR_OBJECT_TRACKER__ASSOCIATION__SOLVER__SSP_HPP_
