+++ perception/radar_object_tracker/package.xml
@@ -27,7 +27,6 @@
   <depend>rclcpp_components</depend>
   <depend>tf2</depend>
   <depend>tf2_ros</depend>
-  <depend>tier4_perception_msgs</depend>
   <depend>unique_identifier_msgs</depend>
   <depend>yaml-cpp</depend>
