+++ perception/radar_object_tracker/src/association/mu_ssp/mu_successive_shortest_path_wrapper.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/radar_object_tracker/association/solver/mu_ssp.hpp"
+#include "autoware_radar_object_tracker/association/solver/mu_ssp.hpp"
 
 #include <mussp/mussp.h>
