+++ perception/radar_object_tracker/src/radar_object_tracker_node.cpp
@@ -14,10 +14,10 @@
 
 #define EIGEN_MPL2_ONLY
 
-#include "autoware/radar_object_tracker/radar_object_tracker_node.hpp"
+#include "radar_object_tracker_node.hpp"
 
-#include "autoware/radar_object_tracker/utils/radar_object_tracker_utils.hpp"
-#include "autoware/radar_object_tracker/utils/utils.hpp"
+#include "autoware_radar_object_tracker/utils/radar_object_tracker_utils.hpp"
+#include "autoware_radar_object_tracker/utils/utils.hpp"
 
 #include <Eigen/Core>
 #include <Eigen/Geometry>
