+++ perception/radar_object_tracker/src/tracker/model/constant_turn_rate_motion_tracker.cpp
@@ -18,12 +18,12 @@
 
 #define EIGEN_MPL2_ONLY
 
-#include "autoware/radar_object_tracker/tracker/model/constant_turn_rate_motion_tracker.hpp"
+#include "autoware_radar_object_tracker/tracker/model/constant_turn_rate_motion_tracker.hpp"
 
-#include "autoware/radar_object_tracker/utils/utils.hpp"
 #include "autoware/universe_utils/geometry/boost_polygon_utils.hpp"
 #include "autoware/universe_utils/math/unit_conversion.hpp"
 #include "autoware/universe_utils/ros/msg_covariance.hpp"
+#include "autoware_radar_object_tracker/utils/utils.hpp"
 
 #include <bits/stdc++.h>
 #include <tf2/LinearMath/Matrix3x3.h>
