+++ perception/radar_object_tracker/src/utils/radar_object_tracker_utils.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/radar_object_tracker/utils/radar_object_tracker_utils.hpp"
+#include "autoware_radar_object_tracker/utils/radar_object_tracker_utils.hpp"
 
 namespace autoware::radar_object_tracker::utils
 {
